resource_service/app/core/security.py
"""
Security utilities for Resource Service
Load public key, verify JWT tokens
"""

from cryptography.hazmat.primitives import serialization
from cryptography.hazmat.backends import default_backend
import jwt
from typing import Optional, Dict, Any
from datetime import datetime
import os

def load_public_key() -> str:
    """Load public key from file"""
    # TODO: Implement loading public key from file
    pass

def verify_jwt_token(token: str) -> Optional[Dict[str, Any]]:
    """Verify JWT token with public key"""
    # TODO: Implement JWT token verification
    pass

def get_current_user_from_token(token: str) -> Optional[Dict[str, Any]]:
    """Extract user info from verified JWT token"""
    # TODO: Implement user extraction from token
    pass

def is_token_expired(payload: Dict[str, Any]) -> bool:
    """Check if token is expired"""
    # TODO: Implement token expiration check
    pass
