resource_service/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
PyJWT==2.8.0

# Cryptography
cryptography>=41.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
